+++ src/box/Inbox.jsx
@@ -12,24 +12,19 @@ import UploadIcon from 'material-ui/svg-icons/file/cloud-upload'
 import DialogOverlay from '../common/DialogOverlay'
 import FlatButton from '../common/FlatButton'
 import { parseTime } from '../common/datetime'
-import Thumb from '../file/Thumb'
+import MediaBox from './MediaBox'
 
 const curve = 'all 450ms cubic-bezier(0.23, 1, 0.32, 1) 0ms'
 
 const imgUrl = 'http://cn.bing.com/th?id=ABT1B401B62BAA3194420276E294380581BC45A4292AE1FF991F97E75ED74A511A1&w=608&h=200&c=2&rs=1&pid=SANGAM'
 
-const calcCurve = (tsd, wd) => {
-  return `all 450ms cubic-bezier(0.23, 1, 0.32, 1) ${tsd || '0ms'},
-          margin-left 450ms cubic-bezier(0.23, 1, 0.32, 1) ${wd || '0ms'},
-          width 450ms cubic-bezier(0.23, 1, 0.32, 1) ${wd || '0ms'},
-          height 450ms cubic-bezier(0.23, 1, 0.32, 1) ${wd || '0ms'}`
-}
 class Inbox extends React.Component {
   constructor(props) {
     super(props)
 
     this.state = {
-      selected: -1
+      selected: -1,
+      hover: -1
     }
 
     this.handleResize = () => this.forceUpdate()
@@ -87,7 +82,7 @@ class Inbox extends React.Component {
       const selected = i === s
       const diff = h1 - h2
 
-      const height = selected ? h * 1.2 : h
+      const height = selected ? Math.min(h * 2, 562) : h
 
       if (h1 <= h2) { // left
         top = h1
@@ -157,124 +152,6 @@ class Inbox extends React.Component {
     return res // ({ height, content })
   }
 
-  renderData(data) {
-    return (
-      <div style={{ width: 810, position: 'relative' }} >
-        {
-          this.calcPos(this.process(data), this.state.selected).map((v, i) => {
-            const { height, top, left, selected, tsd, wd, content } = v
-            const { type, comment, index, tweeter, list, uuid, box, ctime } = content
-            const isMedia = list && list.every(l => l.metadata)
-            const isMany = list && list.length > 6
-            return (
-              <Paper
-                key={uuid}
-                style={{
-                  height,
-                  position: 'absolute',
-                  backgroundColor: '#FFF',
-                  width: selected ? 750 : 360,
-                  transition: calcCurve(tsd, wd),
-                  margin: `${top}px 15px 0px ${left}px`
-                }}
-                onTouchTap={() => this.handleSelect(i)}
-              >
-                <div
-                  style={{
-                    height: 72,
-                    width: '100%',
-                    display: 'flex',
-                    alignItems: 'center',
-                    transition: calcCurve(tsd, wd)
-                  }}
-                >
-                  <div style={{ width: 16 }} />
-                  <div style={{ borderRadius: 20, width: 40, height: 40, overflow: 'hidden' }}>
-                    <img width={40} height={40} alt="face" src={imgUrl} />
-                  </div>
-                  <div style={{ width: 16 }} />
-                  <div style={{ flexGrow: 1 }} >
-                    <div style={{ height: 12 }} />
-                    <div style={{ height: 24, fontWeight: 500, display: 'flex', alignItems: 'center' }} >
-                      <div style={{ maxWidth: 216, overflow: 'hidden', whiteSpace: 'nowrap', textOverflow: 'ellipsis' }}>
-                        { `用户-${tweeter.id.slice(0, 4)}` }
-                      </div>
-                      <div style={{ flexGrow: 1 }} />
-                      <div style={{ fontSize: 12, color: 'rgba(0,0,0,.54)' }}>
-                        { parseTime(ctime) }
-                      </div>
-                      <div style={{ width: 24 }} />
-                    </div>
-                    <div style={{ height: 20, fontSize: 14, color: 'rgba(0,0,0,.54)' }} >
-                      { `来自"${box.name || '未命名'}"群，分享了${list.length}${isMedia ? '张照片' : '个文件'}` }
-                    </div>
-                    <div style={{ height: 16 }} />
-                  </div>
-                </div>
-                <div
-                  style={{
-                    width: '100%',
-                    display: 'flex',
-                    overflow: 'hidden',
-                    height: height - 72,
-                    alignItems: 'center',
-                    transition: calcCurve(tsd, wd)
-                  }}
-                >
-                  {
-                    isMedia ? list.map((l, i) => {
-                      const { sha256 } = l
-                      return (
-                        <Thumb
-                          key={sha256 + i}
-                          digest={sha256}
-                          boxUUID={box.uuid}
-                          ipcRenderer={this.props.ipcRenderer}
-                          height={height - 88}
-                          width={selected ? 750 : 360}
-                          full={selected}
-                          imgStyle={{ objectFit: 'cover', transition: calcCurve(tsd, wd) }}
-                        />
-                      )
-                    })
-                    :
-                      <div
-                        style={{
-                          height: 72,
-                          display: 'flex',
-                          width: '100%',
-                          alignItems: 'center',
-                          backgroundColor: '#FFA000',
-                          color: '#FFF'
-                        }}
-                      >
-                        <div style={{ width: 24, height: 24, display: 'flex', alignItems: 'center', padding: 16 }}>
-                          <FileFolder color="#FFF" />
-                        </div>
-                        <div
-                          style={{
-                            maxWidth: list.length === 1 ? 300 : 120,
-                            overflow: 'hidden',
-                            whiteSpace: 'nowrap',
-                            textOverflow: 'ellipsis'
-                          }}
-                        >
-                          { list[0].filename }
-                        </div>
-                        <div style={{ width: 4 }} />
-                        { list.length > 1 && i18n.__n('And Other %s Items', list.length)}
-                      </div>
-                  }
-                </div>
-              </Paper>
-            )
-          })
-        }
-        <div style={{ height: 24 }} />
-      </div>
-    )
-  }
-
   render() {
     console.log('Box', this.props, this.state)
     return (
@@ -307,7 +184,21 @@ class Inbox extends React.Component {
              >
                <CircularProgress />
              </div> :
-            this.props.tweets.length ? this.renderData(this.props.tweets) : this.renderNoData()
+            this.props.tweets.length ?
+            <div style={{ width: 810, position: 'relative' }}>
+              {
+                this.calcPos(this.process(this.props.tweets), this.state.selected).map((v, i) => (
+                  <MediaBox
+                    key={v.content.uuid}
+                    i={i}
+                    data={v}
+                    handleSelect={this.handleSelect}
+                    ipcRenderer={this.props.ipcRenderer}
+                  />
+                ))
+              }
+            </div>
+            : this.renderNoData()
         }
 
         {/* Selected Header */}
